requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
networkx>=3.1
tqdm>=4.65.0
gensim>=4.3.0
wandb>=0.15.0
jupyter>=1.0.0
scikit-learn>=1.3.0
PyYAML>=6.0.0
pytest>=7.4.0
black>=23.0.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.5.0
pre-commit>=3.3.0
notebook>=7.0.0
ipywidgets>=8.0.0
plotly>=5.15.0